Add MCP2515 receive buffer 0 rollover enable
The MCP2515 has two receive buffers.  The first is the higher priority of the two,
and the chip can be configured to store the next frame in the second buffer if
the first is full and hasn't yet been read.  A function to enable/ disable this
rollover (called BUKT in the datasheet) has been added.